metrics/sender/requirements.txt
py-dateutil
requests
